--- data/duke.txt
@@ -1,2 +1,0 @@
-T | 0 | test
-T | 0 | test 2
