+++ src/main/java/duke/tasks/Deadline.java
@@ -1,45 +1,43 @@
-package tasks;
+package duke.tasks;
 
 import exception.DukeException;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
 
-public class Event extends Task {
+public class Deadline extends Task {
 
-    protected LocalDate start;
-    protected LocalDate end;
+    protected LocalDate by;
 
-    public Event(String description, String start, String end) throws DukeException {
+    public Deadline(String description, String by) throws DukeException {
         super(description);
         try {
-            this.start =  LocalDate.parse(start);
-            this.end =  LocalDate.parse(end);
+            this.by = LocalDate.parse(by);
         } catch (DateTimeParseException e) {
             throw new DukeException("Parse Error: " + e.getMessage() +  "\n" +
                     "\tAccepted format: \"YYYY-MM-DD\"");
         }
-
     }
 
     @Override
     public String getData() {
         StringBuilder sb = new StringBuilder();
-        sb.append("E | ");
+        sb.append("D | ");
         if (this.isDone) {
             sb.append("1 | ");
         } else {
             sb.append("0 | ");
         }
         sb.append(this.description).append(" | ");
-        sb.append(this.start).append(" | ").append(this.end).append("\n");;
+        sb.append(this.by).append("\n");;
         return sb.toString();
     }
 
     @Override
     public String toString() {
-        return "[E]" + super.toString() +
-        " (from: " + start.getMonth().toString().substring(0, 3) + " " + start.getDayOfMonth() + " " + start.getYear() +
-        " to: " + end.getMonth().toString().substring(0, 3) + " " + end.getDayOfMonth() + " " + end.getYear() + ")";
+        return "[D]" + super.toString() + " (by: " +
+                by.getMonth().toString().substring(0, 3) + " " +
+                by.getDayOfMonth() + " " +
+                by.getYear() + ")";
     }
 }
