+++ src/main/java/duke/tasks/Event.java
@@ -1,43 +1,45 @@
-package tasks;
+package duke.tasks;
 
 import exception.DukeException;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
 
-public class Deadline extends Task {
+public class Event extends Task {
 
-    protected LocalDate by;
+    protected LocalDate start;
+    protected LocalDate end;
 
-    public Deadline(String description, String by) throws DukeException {
+    public Event(String description, String start, String end) throws DukeException {
         super(description);
         try {
-            this.by = LocalDate.parse(by);
+            this.start =  LocalDate.parse(start);
+            this.end =  LocalDate.parse(end);
         } catch (DateTimeParseException e) {
             throw new DukeException("Parse Error: " + e.getMessage() +  "\n" +
                     "\tAccepted format: \"YYYY-MM-DD\"");
         }
+
     }
 
     @Override
     public String getData() {
         StringBuilder sb = new StringBuilder();
-        sb.append("D | ");
+        sb.append("E | ");
         if (this.isDone) {
             sb.append("1 | ");
         } else {
             sb.append("0 | ");
         }
         sb.append(this.description).append(" | ");
-        sb.append(this.by).append("\n");;
+        sb.append(this.start).append(" | ").append(this.end).append("\n");;
         return sb.toString();
     }
 
     @Override
     public String toString() {
-        return "[D]" + super.toString() + " (by: " +
-                by.getMonth().toString().substring(0, 3) + " " +
-                by.getDayOfMonth() + " " +
-                by.getYear() + ")";
+        return "[E]" + super.toString() +
+        " (from: " + start.getMonth().toString().substring(0, 3) + " " + start.getDayOfMonth() + " " + start.getYear() +
+        " to: " + end.getMonth().toString().substring(0, 3) + " " + end.getDayOfMonth() + " " + end.getYear() + ")";
     }
 }
