+++ src/main/java/exception/DukeException.java
@@ -8,6 +8,6 @@ public class DukeException extends Exception {
 
     @Override
     public String toString() {
-        return LINES + "\t☹ OOPS!!! " + getMessage() + "\n" + LINES;
+        return "☹ OOPS!!! " + getMessage();
     }
 }
