+++ text-ui-test/EXPECTED.TXT
@@ -6,11 +6,11 @@
 	____________________________________________________________
 	Got it. I've added this task:
 	  [T][ ] borrow book
-	Now you have 1 tasks in the list.
+	Now you have 1 duke.tasks in the list.
 	____________________________________________________________
 
 	____________________________________________________________
-	Here are the tasks in your list:
+	Here are the duke.tasks in your list:
 	1. [T][ ] borrow book
 	____________________________________________________________
 	____________________________________________________________
